+++ authentication/authentication-service-rust-crate/src/app.rs
@@ -56,6 +56,7 @@ impl Hooks for App {
         AppRoutes::with_default_routes() // controller routes below
             .add_route(controllers::auth::routes())
             .add_route(controllers::jwks::routes())
+            .add_route(controllers::docs::routes())
     }
     async fn connect_workers(ctx: &AppContext, queue: &Queue) -> Result<()> {
         queue.register(DownloadWorker::build(ctx)).await?;