+++ case-folder/case-folder-front-end-with-svelte/svelte.config.js
@@ -2,7 +2,7 @@ import adapter from '@sveltejs/adapter-auto';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 const LILY_HELPERS =
-    '../../../lilydesignsystem/lily-design-system/lily-design-system-svelte-helpers';
+    '../../../../lilydesignsystem/lily-design-system/lily-design-system-svelte-helpers';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {